Tests: Cypress tests imrpovements (#2816)
* Update tests structure
* Increase containers for parallel execution
* Update workflows
* Add test ids
