+++ .github/workflows/smoke.yml
@@ -1,8 +1,7 @@
-name: Safe Apps e2e
+name: Smoke tests
 
 on:
   pull_request:
-  workflow_dispatch:
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
@@ -10,12 +9,12 @@ concurrency:
 
 jobs:
   e2e:
-    runs-on: ubuntu-latest
-    name: Cypress Safe Apps tests
+    runs-on: ubuntu-20.04
+    name: Cypress Smoke tests
     strategy:
       fail-fast: false
       matrix:
-        containers: [1, 2]
+        containers: [1, 2, 3, 4, 5]
     steps:
       - uses: actions/checkout@v3
 
@@ -36,12 +35,11 @@ jobs:
       - uses: cypress-io/github-action@v4
         with:
           parallel: true
-          spec: cypress/e2e/safe-apps/*.cy.js
+          spec: cypress/e2e/smoke/*.cy.js
           browser: chrome
           record: true
           config: baseUrl=http://localhost:8080
-          group: 'Safe Apps tests'
+          group: 'Smoke tests'
         env:
-          CYPRESS_PROJECT_ID: okn21k
-          CYPRESS_RECORD_KEY: ${{ secrets.CYPRESS_SAFE_APPS_RECORD_KEY }}
+          CYPRESS_RECORD_KEY: ${{ secrets.CYPRESS_RECORD_KEY }}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
