+++ cypress/e2e/pages/main.page.js
@@ -11,6 +11,16 @@ export function clickOnSideMenuItem(item) {
   cy.get('p').contains(item).click()
 }
 
+export function waitForTrnsactionHistoryToComplete() {
+  cy.intercept('GET', constants.transactionHistoryEndpoint).as('History')
+  cy.wait('@History')
+}
+
+export function waitForSafeListRequestToComplete() {
+  cy.intercept('GET', constants.safeListEndpoint).as('Safes')
+  cy.wait('@Safes')
+}
+
 export function acceptCookies(index = 0) {
   cy.wait(1000)
