+++ cypress/e2e/pages/owners.pages.js
@@ -1,11 +1,12 @@
 import * as constants from '../../support/constants'
 import * as main from '../pages/main.page'
+import * as createWallet from '../pages/create_wallet.pages'
+import * as navigation from '../pages/navigation.page'
 
 const copyToClipboardBtn = 'button[aria-label="Copy to clipboard"]'
 const tooltipLabel = (label) => `span[aria-label="${label}"]`
 const removeOwnerBtn = 'span[data-track="settings: Remove owner"] > span > button'
 const replaceOwnerBtn = 'span[data-track="settings: Replace owner"] > span > button'
-const addOwnerBtn = 'span[data-track="settings: Add owner"] > button'
 const tooltip = 'div[role="tooltip"]'
 const expandMoreIcon = 'svg[data-testid="ExpandMoreIcon"]'
 const sentinelStart = 'div[data-testid="sentinelStart"]'
@@ -21,12 +22,13 @@ const existingOwnerAddressInput = (index) => `input[name="owners.${index}.addres
 const existingOwnerNameInput = (index) => `input[name="owners.${index}.name"]`
 const singleOwnerNameInput = 'input[name="name"]'
 const finishTransactionBtn = '[data-testid="finish-transaction-btn"]'
+const addOwnerBtn = '[data-testid="add-owner-btn"]'
+const addOwnerNextBtn = '[data-testid="add-owner-next-btn"]'
 
 const disconnectBtnStr = 'Disconnect'
 const notConnectedStatus = 'Connect'
 const e2eWalletStr = 'E2E Wallet'
 const max50charsLimitStr = 'Maximum 50 symbols'
-const nextBtnStr = 'Next'
 const executeBtnStr = 'Execute'
 const backbtnStr = 'Back'
 const removeOwnerStr = 'Remove owner'
@@ -160,12 +162,11 @@ export function clickOnConnectBtn() {
 }
 
 export function waitForConnectionStatus() {
-  cy.get('div').contains(e2eWalletStr)
+  cy.get(createWallet.accountInfoHeader).should('exist')
 }
 
 export function openAddOwnerWindow() {
-  cy.get('span').contains(addNewOwnerStr).click()
-  cy.wait(1000)
+  cy.get(addOwnerBtn).should('be.enabled').click()
   cy.get(newOwnerName).should('be.visible')
   cy.get(newOwnerAddress).should('be.visible')
 }
@@ -203,11 +204,11 @@ export function verifyNewOwnerName(name) {
 }
 
 export function clickOnNextBtn() {
-  cy.get('button').contains(nextBtnStr).click()
+  cy.get(addOwnerNextBtn).should('be.enabled').click()
 }
 
 export function clickOnBackBtn() {
-  cy.get('button').contains(backbtnStr).click()
+  cy.get(navigation.modalBackBtn).should('be.enabled').click()
 }
 
 export function verifyConfirmTransactionWindowDisplayed() {
