+++ cypress/e2e/regression/add_owner.cy.js
@@ -1,0 +1,55 @@
+import * as constants from '../../support/constants'
+import * as main from '../../e2e/pages/main.page'
+import * as owner from '../pages/owners.pages'
+import * as addressBook from '../pages/address_book.page'
+
+describe('Add Owners tests', () => {
+  beforeEach(() => {
+    cy.visit(constants.setupUrl + constants.SEPOLIA_TEST_SAFE_1)
+    cy.clearLocalStorage()
+    main.acceptCookies()
+    cy.contains(owner.safeAccountNonceStr, { timeout: 10000 })
+  })
+
+  it('Verify Tooltip displays correct message for disconnected user', () => {
+    owner.waitForConnectionStatus()
+    owner.clickOnWalletExpandMoreIcon()
+    owner.clickOnDisconnectBtn()
+    owner.verifyAddOwnerBtnIsDisabled()
+  })
+
+  it('Verify the Add New Owner Form can be opened', () => {
+    owner.waitForConnectionStatus()
+    owner.openAddOwnerWindow()
+  })
+
+  it('Verify error message displayed if character limit is exceeded in Name input', () => {
+    owner.waitForConnectionStatus()
+    owner.openAddOwnerWindow()
+    owner.typeOwnerName(main.generateRandomString(51))
+    owner.verifyErrorMsgInvalidAddress(constants.addressBookErrrMsg.exceedChars)
+  })
+
+  it('Verify that the "Name" field is auto-filled with the relevant name from Address Book', () => {
+    cy.visit(constants.addressBookUrl + constants.SEPOLIA_TEST_SAFE_1)
+    addressBook.clickOnCreateEntryBtn()
+    addressBook.typeInName(constants.addresBookContacts.user1.name)
+    addressBook.typeInAddress(constants.addresBookContacts.user1.address)
+    addressBook.clickOnSaveEntryBtn()
+    addressBook.verifyNewEntryAdded(constants.addresBookContacts.user1.name, constants.addresBookContacts.user1.address)
+    cy.visit(constants.setupUrl + constants.SEPOLIA_TEST_SAFE_1)
+    owner.waitForConnectionStatus()
+    owner.openAddOwnerWindow()
+    owner.typeOwnerAddress(constants.addresBookContacts.user1.address)
+    owner.selectNewOwner(constants.addresBookContacts.user1.name)
+    owner.verifyNewOwnerName(constants.addresBookContacts.user1.name)
+  })
+
+  it('Verify that Name field not mandatory', () => {
+    owner.waitForConnectionStatus()
+    owner.openAddOwnerWindow()
+    owner.typeOwnerAddress(constants.SEPOLIA_OWNER_2)
+    owner.clickOnNextBtn()
+    owner.verifyConfirmTransactionWindowDisplayed()
+  })
+})
