+++ cypress/e2e/regression/balances_pagination.cy.js
