+++ cypress/e2e/regression/beamer.cy.js
