+++ cypress/e2e/regression/pending_actions.cy.js
