+++ cypress/e2e/smoke/add_owner.cy.js
@@ -1,68 +1,29 @@
 import * as constants from '../../support/constants'
 import * as main from '../../e2e/pages/main.page'
 import * as owner from '../pages/owners.pages'
-import * as addressBook from '../pages/address_book.page'
+import * as navigation from '../pages/navigation.page'
 
-describe('Add Owners tests', () => {
+describe('[SMOKE] Add Owners tests', () => {
   beforeEach(() => {
     cy.visit(constants.setupUrl + constants.SEPOLIA_TEST_SAFE_1)
     cy.clearLocalStorage()
+    main.waitForTrnsactionHistoryToComplete()
     main.acceptCookies()
-    cy.contains(owner.safeAccountNonceStr, { timeout: 10000 })
+    main.verifyElementsExist([navigation.setupSection])
   })
 
-  it('Verify the presence of "Add Owner" button', () => {
+  it('[SMOKE] Verify the presence of "Add Owner" button', () => {
     owner.verifyAddOwnerBtnIsEnabled()
   })
 
-  it('Verify “Add new owner” button tooltip displays correct message for Non-Owner', () => {
+  it('[SMOKE] Verify “Add new owner” button tooltip displays correct message for Non-Owner', () => {
     cy.visit(constants.setupUrl + constants.SEPOLIA_TEST_SAFE_2)
+    main.waitForTrnsactionHistoryToComplete()
     owner.verifyAddOwnerBtnIsDisabled()
   })
 
-  it('Verify Tooltip displays correct message for disconnected user', () => {
-    owner.waitForConnectionStatus()
-    owner.clickOnWalletExpandMoreIcon()
-    owner.clickOnDisconnectBtn()
-    owner.verifyAddOwnerBtnIsDisabled()
-  })
-
-  it('Verify the Add New Owner Form can be opened', () => {
-    owner.waitForConnectionStatus()
-    owner.openAddOwnerWindow()
-  })
-
-  it('Verify error message displayed if character limit is exceeded in Name input', () => {
-    owner.waitForConnectionStatus()
-    owner.openAddOwnerWindow()
-    owner.typeOwnerName(main.generateRandomString(51))
-    owner.verifyErrorMsgInvalidAddress(constants.addressBookErrrMsg.exceedChars)
-  })
-
-  it('Verify that the "Name" field is auto-filled with the relevant name from Address Book', () => {
-    cy.visit(constants.addressBookUrl + constants.SEPOLIA_TEST_SAFE_1)
-    addressBook.clickOnCreateEntryBtn()
-    addressBook.typeInName(constants.addresBookContacts.user1.name)
-    addressBook.typeInAddress(constants.addresBookContacts.user1.address)
-    addressBook.clickOnSaveEntryBtn()
-    addressBook.verifyNewEntryAdded(constants.addresBookContacts.user1.name, constants.addresBookContacts.user1.address)
-    cy.visit(constants.setupUrl + constants.SEPOLIA_TEST_SAFE_1)
-    owner.waitForConnectionStatus()
-    owner.openAddOwnerWindow()
-    owner.typeOwnerAddress(constants.addresBookContacts.user1.address)
-    owner.selectNewOwner(constants.addresBookContacts.user1.name)
-    owner.verifyNewOwnerName(constants.addresBookContacts.user1.name)
-  })
-
-  it('Verify that Name field not mandatory', () => {
-    owner.waitForConnectionStatus()
-    owner.openAddOwnerWindow()
-    owner.typeOwnerAddress(constants.SEPOLIA_OWNER_2)
-    owner.clickOnNextBtn()
-    owner.verifyConfirmTransactionWindowDisplayed()
-  })
-
-  it('Verify relevant error messages are displayed in Address input', () => {
+  // TODO: Check if this test is covered with unit tests
+  it('[SMOKE] Verify relevant error messages are displayed in Address input', () => {
     owner.waitForConnectionStatus()
     owner.openAddOwnerWindow()
     owner.typeOwnerAddress(main.generateRandomString(10))
@@ -81,22 +42,20 @@ describe('Add Owners tests', () => {
     owner.verifyErrorMsgInvalidAddress(constants.addressBookErrrMsg.alreadyAdded)
   })
 
-  it('Verify default threshold value. Verify correct threshold calculation', () => {
+  it('[SMOKE] Verify default threshold value. Verify correct threshold calculation', () => {
     owner.waitForConnectionStatus()
     owner.openAddOwnerWindow()
     owner.typeOwnerAddress(constants.DEFAULT_OWNER_ADDRESS)
     owner.verifyThreshold(1, 2)
   })
 
-  it('Verify valid Address validation', () => {
+  it('[SMOKE] Verify valid Address validation', () => {
     owner.waitForConnectionStatus()
     owner.openAddOwnerWindow()
     owner.typeOwnerAddress(constants.SEPOLIA_OWNER_2)
     owner.clickOnNextBtn()
     owner.verifyConfirmTransactionWindowDisplayed()
-    cy.reload()
-    owner.waitForConnectionStatus()
-    owner.openAddOwnerWindow()
+    owner.clickOnBackBtn()
     owner.typeOwnerAddress(constants.SEPOLIA_TEST_SAFE_2)
     owner.clickOnNextBtn()
     owner.verifyConfirmTransactionWindowDisplayed()
