+++ cypress/e2e/smoke/batch_tx.cy.js
@@ -7,7 +7,7 @@ const currentNonce = 3
 const funds_first_tx = '0.001'
 const funds_second_tx = '0.002'
 
-describe('Batch transaction tests', () => {
+describe('[SMOKE] Batch transaction tests', () => {
   beforeEach(() => {
     cy.clearLocalStorage()
     cy.visit(constants.BALANCE_URL + constants.SEPOLIA_TEST_SAFE_5)
@@ -15,17 +15,12 @@ describe('Batch transaction tests', () => {
     main.acceptCookies()
   })
 
-  it('Verify empty batch list can be opened', () => {
+  it('[SMOKE] Verify empty batch list can be opened', () => {
     batch.openBatchtransactionsModal()
     batch.openNewTransactionModal()
   })
 
-  it('Verify the Add batch button is present in a transaction form', () => {
-    //The "true" is to validate that the add to batch button is not visible if "Yes, execute" is selected
-    batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
-  })
-
-  it('Verify a transaction can be added to the batch', () => {
+  it('[SMOKE] Verify a transaction can be added to the batch', () => {
     batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
     cy.contains(batch.transactionAddedToBatchStr).should('be.visible')
     //The batch button in the header shows the transaction count
@@ -34,16 +29,7 @@ describe('Batch transaction tests', () => {
     batch.verifyAmountTransactionsInBatch(1)
   })
 
-  it('Verify a second transaction can be added to the batch', () => {
-    batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
-    cy.wait(1000)
-    batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
-    batch.verifyBatchIconCount(2)
-    batch.clickOnBatchCounter()
-    batch.verifyAmountTransactionsInBatch(2)
-  })
-
-  it('Verify the batch can be confirmed and related transactions exist in the form', () => {
+  it('[SMOKE] Verify the batch can be confirmed and related transactions exist in the form', () => {
     batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
     cy.wait(1000)
     batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
@@ -58,7 +44,7 @@ describe('Batch transaction tests', () => {
     cy.get('@TransactionList').find('li').eq(1).find('span').eq(0).contains(funds_first_tx)
   })
 
-  it('Verify a transaction can be removed from the batch', () => {
+  it('[SMOKE] Verify a transaction can be removed from the batch', () => {
     batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
     cy.wait(1000)
     batch.addNewTransactionToBatch(constants.EOA, currentNonce, funds_first_tx)
