+++ cypress/e2e/smoke/landing.cy.js
@@ -1,6 +1,6 @@
 import * as constants from '../../support/constants'
-describe('Landing page tests', () => {
-  it('Verify a user will be redirected to welcome page', () => {
+describe('[SMOKE] Landing page tests', () => {
+  it('[SMOKE] Verify a user will be redirected to welcome page', () => {
     cy.clearLocalStorage()
     cy.visit('/')
     cy.url().should('include', constants.welcomeUrl)
