+++ src/components/new-safe/create/steps/OwnerPolicyStep/index.tsx
@@ -170,10 +170,16 @@ const OwnerPolicyStep = ({
         <Divider />
         <Box className={layoutCss.row}>
           <Box display="flex" flexDirection="row" justifyContent="space-between" gap={3}>
-            <Button variant="outlined" size="small" onClick={handleBack} startIcon={<ArrowBackIcon fontSize="small" />}>
+            <Button
+              data-testid="back-btn"
+              variant="outlined"
+              size="small"
+              onClick={handleBack}
+              startIcon={<ArrowBackIcon fontSize="small" />}
+            >
               Back
             </Button>
-            <Button type="submit" variant="contained" size="stretched" disabled={isDisabled}>
+            <Button data-testid="next-btn" type="submit" variant="contained" size="stretched" disabled={isDisabled}>
               Next
             </Button>
           </Box>
